requirements.txt
torch==1.6.0
natsort==7.0.1
addict==2.2.0
opencv_python==4.2.0.34
dropblock==0.3.0
numpy==1.19.5
matplotlib==3.2.1
torchvision==0.7.0
Pillow==8.3.2
PyYAML==5.4.1
tensorboardX==2.4
